# lib/api-zod/src/generated/types/createExpenseEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Money Saathi API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExpenseEntryFrequency } from "./createExpenseEntryFrequency";

export interface CreateExpenseEntry {
  category: string;
  amount: number;
  frequency: CreateExpenseEntryFrequency;
  description?: string | null;
}