# lib/api-zod/src/generated/types/createProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Money Saathi API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProfileRequestProfileType } from "./createProfileRequestProfileType";

export interface CreateProfileRequest {
  profileType: CreateProfileRequestProfileType;
  businessName?: string | null;
  currency?: string;
}